removed logs from bottom bar
