+++ android/src/main/java/io/parity/signer/screens/keysetdetails/KeySetDetailsScreenView.kt
@@ -28,7 +28,7 @@ import androidx.compose.ui.unit.dp
 import io.parity.signer.R
 import io.parity.signer.components.base.SecondaryButtonWide
 import io.parity.signer.components.exposesecurity.ExposedIcon
-import io.parity.signer.components.panels.BottomBar2
+import io.parity.signer.components.panels.BottomBar
 import io.parity.signer.components.panels.BottomBar2State
 import io.parity.signer.domain.*
 import io.parity.signer.screens.keysetdetails.items.NetworkKeysExpandable
@@ -108,7 +108,7 @@ fun KeySetDetailsScreenView(
 					.padding(end = 16.dp, bottom = 24.dp)
 			)
 		}
-		BottomBar2(navigator, BottomBar2State.KEYS)
+		BottomBar(navigator, BottomBar2State.KEYS)
 	}
 }
