+++ android/src/main/java/io/parity/signer/screens/settings/backup/SeedBackupScreen.kt
@@ -23,10 +23,9 @@ import androidx.compose.ui.res.dimensionResource
 import androidx.compose.ui.res.stringResource
 import androidx.compose.ui.tooling.preview.Preview
 import androidx.compose.ui.unit.dp
-import androidx.lifecycle.viewmodel.compose.viewModel
 import io.parity.signer.R
 import io.parity.signer.components.base.ScreenHeader
-import io.parity.signer.components.panels.BottomBar2
+import io.parity.signer.components.panels.BottomBar
 import io.parity.signer.components.panels.BottomBar2State
 import io.parity.signer.domain.Callback
 import io.parity.signer.domain.EmptyNavigator
@@ -63,7 +62,7 @@ fun SeedBackupScreen(
 				}
 			}
 		}
-		BottomBar2(
+		BottomBar(
 			rootNavigator, BottomBar2State.SETTINGS,
 			skipRememberCameraParent = true
 		)
