+++ android/src/main/java/io/parity/signer/screens/settings/networks/list/NetworksListScreen.kt
@@ -21,7 +21,7 @@ import androidx.compose.ui.unit.dp
 import io.parity.signer.R
 import io.parity.signer.components.base.ScreenHeader
 import io.parity.signer.components.networkicon.NetworkIcon
-import io.parity.signer.components.panels.BottomBar2
+import io.parity.signer.components.panels.BottomBar
 import io.parity.signer.components.panels.BottomBar2State
 import io.parity.signer.components.panels.CameraParentScreen
 import io.parity.signer.components.panels.CameraParentSingleton
@@ -54,7 +54,7 @@ fun NetworksListScreen(model: NetworksListModel, rootNavigator: Navigator) {
 				}
 			}
 		}
-		BottomBar2(
+		BottomBar(
 			rootNavigator, BottomBar2State.SETTINGS,
 			skipRememberCameraParent = true
 		) {
