--- android/src/main/res/drawable/ic_view_agenda_outlined_24.xml
@@ -1,12 +1,0 @@
-<vector xmlns:android="http://schemas.android.com/apk/res/android"
-    android:width="24dp"
-    android:height="24dp"
-    android:viewportWidth="24"
-    android:viewportHeight="24">
-  <path
-      android:pathData="M19,13H5C3.9,13 3,13.9 3,15V19C3,20.1 3.9,21 5,21H19C20.1,21 21,20.1 21,19V15C21,13.9 20.1,13 19,13ZM19,19H5V15H19V19Z"
-      android:fillColor="#000000"/>
-  <path
-      android:pathData="M19,3H5C3.9,3 3,3.9 3,5V9C3,10.1 3.9,11 5,11H19C20.1,11 21,10.1 21,9V5C21,3.9 20.1,3 19,3ZM19,9H5V5H19V9Z"
-      android:fillColor="#000000"/>
-</vector>
